+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v0.1.0 (unreleased)
+* Add params method to beers.  Allows passing additional parameters in an additional way.
 * Add params method to beer.  Allows passing additional parameters in an additional way.
 
 ## v0.0.1